requirements.txt
aniso8601==9.0.1
bcrypt==3.2.0
certifi==2021.5.30
cffi==1.14.6
click==8.0.1
cloudinary==1.26.0
colorama==0.4.4
cryptography==3.4.8
databases==0.5.2
fernet==1.0.1
Flask==2.0.1
Flask-Cors==3.0.10
Flask-JWT==0.3.2
Flask-RESTful==0.3.9
Flask-SQLAlchemy==2.5.1
flask-swagger-ui==3.36.0
greenlet==1.1.1
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
mysqlclient==2.0.3
orm==0.1.9
psycopg2==2.9.1
pyaes==1.6.1
pycparser==2.20
PyJWT==1.4.2
python-dotenv==0.19.0
pytz==2021.1
six==1.16.0
SQLAlchemy==1.4.23
typesystem==0.3.0
urllib3==1.26.6
Werkzeug==2.0.1
